WIP refactored for knex cleaner and seeded data

+++ utils/data/migrations/20190728163109_001_users.js
@@ -26,7 +26,7 @@ exports.up = function(knex) {
   .createTable('schedule', tbl =>{
     tbl.increments();
 
-    tbl.decimal('plant_id')
+    tbl.float('plant_id')
       .notNullable()
       .references('id')
       .inTable('plants')
